change namespace + thong tin

+++ src/Files/LocalTemporaryFile.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Maatwebsite\Excel\Files;
+namespace Omt\ExcelHelper\Files;
 
 class LocalTemporaryFile extends TemporaryFile
 {
@@ -40,6 +40,10 @@ class LocalTemporaryFile extends TemporaryFile
      */
     public function delete(): bool
     {
+        if (@unlink($this->filePath) || !$this->exists()) {
+            return true;
+        }
+
         return unlink($this->filePath);
     }
